broke one component into its own file for practice, cleaned up changeQuarter function
